changed DTO packaging
