+++ src/main/java/igrn/todo/dto/board/BoardDto.java
@@ -1,5 +1,6 @@
-package igrn.todo.dto;
+package igrn.todo.dto.board;
 
+import igrn.todo.dto.column.ColumnDto;
 import lombok.Getter;
 
 import java.util.List;
