+++ src/main/java/igrn/todo/dto/column/ColumnShortDto.java
@@ -1,13 +1,13 @@
-package igrn.todo.dto;
+package igrn.todo.dto.column;
 
 import lombok.Getter;
 
 @Getter
-public class TicketDto {
+public class ColumnShortDto {
     private final Integer id;
     private final String title;
 
-    public TicketDto(Integer id, String title) {
+    public ColumnShortDto(Integer id, String title) {
         this.id = id;
         this.title = title;
     }
