+++ src/main/java/igrn/todo/dto/column/ColumnTitleDto.java
@@ -1,13 +1,13 @@
-package igrn.todo.dto;
+package igrn.todo.dto.column;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.Getter;
 
 @Getter
-public class BoardTitleDto {
+public class ColumnTitleDto {
     private final String title;
 
-    public BoardTitleDto(@JsonProperty("title") String title) {
+    public ColumnTitleDto(@JsonProperty("title") String title) {
         this.title = title;
     }
 }
