+++ src/main/java/igrn/todo/dto/ticket/TicketDto.java
@@ -1,13 +1,13 @@
-package igrn.todo.dto;
+package igrn.todo.dto.ticket;
 
 import lombok.Getter;
 
 @Getter
-public class ColumnShortDto {
+public class TicketDto {
     private final Integer id;
     private final String title;
 
-    public ColumnShortDto(Integer id, String title) {
+    public TicketDto(Integer id, String title) {
         this.id = id;
         this.title = title;
     }
