+++ src/main/java/igrn/todo/dto/ticket/TicketTitleDto.java
@@ -1,4 +1,4 @@
-package igrn.todo.dto;
+package igrn.todo.dto.ticket;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.Getter;
